+++ src/test/java/school/redrover/AlexLeoGroupTests.java
@@ -1,15 +1,11 @@
 package school.redrover;
 
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import org.apache.commons.lang3.RandomStringUtils;
 import org.openqa.selenium.By;
-import org.openqa.selenium.Keys;
 import org.openqa.selenium.Point;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
-import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.runner.BaseTest;
 
@@ -19,12 +15,6 @@ import java.util.Arrays;
 import java.util.List;
 
 public class AlexLeoGroupTests extends BaseTest {
-
-    private static final String USER_FULL_NAME = RandomStringUtils.randomAlphanumeric(13);
-
-    private static final String DESCRIPTION = RandomStringUtils.randomAlphanumeric(130) + "\n\n" + RandomStringUtils.randomAlphanumeric(23);
-
-    private static final By USER_NAME_LINK = By.xpath("//a[@href='/user/admin']");
 
     @Test
     public void testVerifyLogoJenkinsIsPresent() {
@@ -211,17 +201,6 @@ public class AlexLeoGroupTests extends BaseTest {
         Assert.assertEquals(logoutLink, "log out");
     }
 
-    @Ignore
-    @Test
-    public void testSearchField() {
-        WebElement searchBox = getDriver().findElement(By.id("search-box"));
-        searchBox.sendKeys("");
-        searchBox.sendKeys(Keys.RETURN);
-
-        Assert.assertTrue(getWait5().until(ExpectedConditions.textToBe
-                (By.xpath("//div[@class='jenkins-app-bar__content']/h1"), "Built-In Node")));
-    }
-
     @Test
     public void testNewFreestyleProjectVerification() {
         String nameOfProject = "NewProject2023";
@@ -391,37 +370,6 @@ public class AlexLeoGroupTests extends BaseTest {
         Assert.assertTrue(getDriver().findElement(By.id("jenkins-head-icon")).isDisplayed());
     }
 
-    @Ignore
-    @Test
-    public void testVerifyUserPageMenu() {
-        List<String> listMenuExpected = Arrays.asList("People", "Status", "Builds", "Configure", "My Views", "Credentials");
-
-        getDriver().findElement(By.xpath("//a[@class='model-link']")).click();
-
-        getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.id("description")));
-        List<WebElement> listMenu = getDriver().findElements(By.className("task"));
-
-        for (int i = 0; i < listMenu.size(); i++) {
-            Assert.assertEquals(listMenu.get(i).getText(), listMenuExpected.get(i));
-        }
-    }
-
-    @Ignore
-    @Test
-    public void testVerifyChangeNameUser() {
-        getDriver().findElement(USER_NAME_LINK).click();
-
-        WebElement configure = getWait5().until(ExpectedConditions.elementToBeClickable(By.xpath("//a[@href='/user/admin/configure']")));
-        configure.click();
-
-        WebElement fullName = getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//input[@name='_.fullName']")));
-        fullName.clear();
-        fullName.sendKeys(USER_FULL_NAME);
-        getDriver().findElement(By.xpath("//button[@name='Submit']")).click();
-
-        Assert.assertEquals(getDriver().findElement(USER_NAME_LINK).getText(), USER_FULL_NAME);
-    }
-
     @Test
     public void testTasksMenuNavigation() {
         WebElement newItemElementLink = getDriver().findElement(By.linkText("New Item"));
@@ -452,22 +400,6 @@ public class AlexLeoGroupTests extends BaseTest {
         Point point = new Point(19, 8);
         Assert.assertEquals(findLogoJenkins.getLocation(), point);
 
-    }
-
-    @Ignore
-    @Test
-    public void testVerifyUserDescription() {
-        getDriver().findElement(USER_NAME_LINK).click();
-
-        WebElement editDescription = getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//a[@id='description-link']")));
-        editDescription.click();
-
-        WebElement fullName = getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//textarea[@name='description']")));
-        fullName.clear();
-        fullName.sendKeys(DESCRIPTION);
-        getDriver().findElement(By.xpath("//button[@name='Submit']")).click();
-
-        Assert.assertEquals(getDriver().findElement(By.xpath("//div[@id='description']/div")).getText(), DESCRIPTION);
     }
 
     @Test
